Specify javadoc and move classes
